+++ CHANGELOG.md
@@ -24,6 +24,7 @@ All notable changes of GetSimple Legacy CMS will be documented in this file.
 - New theme function `get_component_title()` to echo or return title of a component.
 - New theme function `get_component_description()` to echo or return description of a component.
 - New theme function `get_page_parent()` to echo or return the slug of the parent page of a current page. This function replace deprecated theme function `get_parent()`.
+- New theme function `get_page_creation_date()` to echo or return page creation date.
 - New theme function `get_site_title()` to echo or return the title of the site. It is an alternative to the function `get_site_name()`.
 - New theme function `get_site_subtile()` to echo or return the subtitle of the site.
 - New theme function `get_site_tagline()` to echo or return the tagline of the site.